fix: add braces to all if statements and split ScriptRunner into smaller files
OxLint CI was failing with "file has too many lines" (476 > 300) and 16x
"expected { after if condition" errors. Split ScriptRunner.tsx into four
focused modules (types, hook, controls, tabs) and added braces to every
braceless if statement. No logic or behavior changes.
